% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveyors.R
\name{add_equation_extensions}
\alias{add_equation_extensions}
\title{Replace array variables in an equation column with dimensional ones}
\usage{
add_equation_extensions(.data, eqn_col, dmx = NA)
}
\arguments{
\item{.data}{input dataframe with all equations to replace}

\item{eqn_col}{column to modify}

\item{dmx}{(optional) external dataframe with name and ext for all variables}
}
\value{
.data with variables in the equation column replaced
}
\description{
Input dataframe has name, ext, and an equation column.  Replace all array
variables in the equation with the appropriate dimensional variable, based
on the extension in the ext column.
}
\details{
Example transformation

Specify dmx if all variables are not represented in .data, or if the variable
names in .data have already been extended.
}
