% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_model_components.R
\name{add_non_negative_check}
\alias{add_non_negative_check}
\title{Check that stocks which are supposed to be non-negative remain so}
\usage{
add_non_negative_check(stock_obj, nn_stocks)
}
\arguments{
\item{stock_obj}{List with `name` and `equation` describing one stock}

\item{nn_stocks}{Vector of non negative stock names}
}
\value{
equation wrapped in a check for negativity
}
\description{
This will just alert us if we have a problem.  To actually ensure
that stocks remain non-negative requires wrapping all outflows.  Currently
there does not seem to be a need for that so I'm putting it off.
}
