% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveyors.R
\name{conveyor_expand_init}
\alias{conveyor_expand_init}
\title{Calculate initial value in slats, and paste with space-separators}
\usage{
conveyor_expand_init(init, len, ext, dt)
}
\arguments{
\item{init}{numerical initial value}

\item{len}{numerical length of conveyor}

\item{ext}{dimension based node name extension}

\item{dt}{time delta}
}
\value{
space-separated initial values for all slats
}
\description{
This is super complicated.  We have the building blocks to calculate, but
for now I'm going to just hardcode the leakage fractions for the one
inital-value conveyor we have.  For this guy we have two leaks.
}
\details{
## General Note On Initial Values

In Stella, initial values are NOT evenly spread among the slots.  They are
placed in such a way that they emulate a steady state (a constant outflow),
given the existing leakages.

Call SN the initial value in slat N (in 1...M), and LF be a single leak
fraction.  With a steady inflow of INF, the value in slot SN would be

SN = INF (1-LF)^(M-N)

Here I take the value BEFORE leakage is removed because leakage is calculated
first before outflow.  Also M is transit time / DT.

We initialize the conveyor with value init = sum_N(SN)
init = sum(INF (1-LF)^(M-N))
     = INF sum((1-LF)^(M-N)
 INF = init/sum((1-LF)^(M-N))

Which gives us our initial vector.


With 2 leaks, we have INF(1-LF1) left after the first leak, then
INF(1-LF1)(1-LF2) after the second.  Let R1 = 1-LF1 and R2 = 1-LF2,
then
Si = (R1 R2)^(M-i)/sum((R1 R2)^(M-N))

## Our hardcoded init

Here we just hardcode initial slat values for our one nonzero initialized
conveyor.  This conveyor has two leaks.  The first has fraction based on
dimension ("Mild" -> 0.0188, else 0.075).  The second is time varying, but
starts at X-0.000001*17.8, where X = .00789 for Severe, .0000667 for
Moderate, and 0 otherwise.  Noting that leakage can't be negative, we have:

                LF1     RM1       LF2     RM2         RM1 RM2
Asymptomatic   .075    .925        0       1            .925
Mild           .0188   .812        0       1            .812
Moderate       .075    .925   .0000489  .9999155    .9249218
Severe         .075    .925   .0078722  .9920922    .9176853

## Other nonzero conveyor inits

For testing purposes, I also put a hardcoded 50% leakage for any other
nonzero conveyor initial values.
}
