% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveyors.R
\name{conveyor_flow_sum}
\alias{conveyor_flow_sum}
\title{Generate net flow equation for a slat based on flow map}
\usage{
conveyor_flow_sum(conveyor_name, ext, idx, nslats, slat_flow, cvy_info)
}
\arguments{
\item{conveyor_name}{Sanitized name of conveyor stock}

\item{ext}{Extension based on dimensions}

\item{idx}{Index of this slat within the conveyor}

\item{nslats}{Number of slats in conveyor (len field divided by DT)}

\item{slat_flow}{Name of inter-slat flow, WITH extension}

\item{cvy_info}{List of inflows, outflows, and leaks WITHOUT extension!}
}
\description{
NOTE: we introduce inter-slat flows which are indexed according to the
      slat they flow INTO.
}
\details{
- Inflows
  1. For the highest index (first) slat, start with any non-leaked inflows
     - NOTE: if a flow is a leak, but does not spread to downstream conveyor,
             we have to sum over the leakage slats!
  2. For all OTHER slats, start with the inter-slat flow with this index
  2. For all slats, ADD any leaked inflows
- Outflows
  1. For all slats, subtract outflow leaks.
  2. For all EXCEPT the last slat, subtract the LOWER index inter-slat flow
  3. for the last slat (index 1), subtract the outflow.

NOTE: leakages AND inter-slat flows are per slat and indexed by slat number

WARNING: I'm making an assumption here that dimensional conveyors are never
         attached to non-dimensional flows or visa-versa.  Just assuming
         extension is the same, which is true for RCM model.
}
