% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveyors.R
\name{conveyor_outflow_leak_effects}
\alias{conveyor_outflow_leak_effects}
\title{Collect leaks which will have an effect on the conveyor outflow}
\usage{
conveyor_outflow_leak_effects(cvy_parms)
}
\arguments{
\item{cvy_parms}{list of useful parameters}
}
\value{
tibble with `flows_from` and `leak_eqn` fields
}
\description{
Return fields:
- flows_from: name of upstream conveyor with dimensional extension
- leak_eqn: equation for the (scalar) value leak from slat 1
}
