% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveyors.R
\name{conveyor_outflow_variables}
\alias{conveyor_outflow_variables}
\title{Generate variable to equation map for flows out of conveyors}
\usage{
conveyor_outflow_variables(cvy_parms)
}
\arguments{
\item{cvy_parms}{list of useful parameters}
}
\value{
tibble with `name` and `equation` for leaks and outflows
}
\description{
Unlike other stocks and flows, conveyors dictate the pace of flow.  Here
we generate equations for all of the flows out of conveyors.  In each case,
we have a "name" field from the sanitized element name, an "ext" field with
the dimension extension, and an equation which dictates the logical flow.
}
\details{
- leaks: each leak is a vector of outflow variables, one for each slat
  - named "name_extension_vec", and then summed into "name_extension" for
    scalar use
- "normal" outflows: final drain from slat 1 after all other external flows
  - start with value stored in slot 1 from previous dt
  - subtract outflow leaks
  - inflow from slot 2 is NOT part of this
}
