% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveyors.R
\name{conveyor_slat_flow_call}
\alias{conveyor_slat_flow_call}
\title{Generate equation string for conveyor slat flows}
\usage{
conveyor_slat_flow_call(base_name, ext, cvy_info)
}
\arguments{
\item{base_name}{Name of conveyor WITHOUT extension or index!}

\item{ext}{Extension based on dimensions}

\item{cvy_info}{data frames for conveyors and flows}
}
\description{
Generate equation string for conveyor slat flows
}
