% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays.R
\name{dimension_extensions}
\alias{dimension_extensions}
\title{Simple map of sanitized name to all extensions (enumerated dimensions)}
\usage{
dimension_extensions(dims_obj)
}
\arguments{
\item{dims_obj}{original dimensions object (will generate `dims_tibble`)}
}
\value{
tibble with `name` and `ext` (dimension-based name extension)
}
\description{
Start by generating a table (vdims) with columns:
- name: the sanitized name of one variable (stock, flow, aux)
- dim.1: the name of the first dimension
- elem.1: name of the element for dimension 1
- ...
}
\details{
There are enough dimension.n and elem.n columns to cover the highest-
dimensional variable.  There is one row for each combination of dimensional
elements for each variable (same order as found in the xml file).

Next we unite all the elem.* columns with an "_" separator into the "ext"
column
}
