% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveyors.R
\name{expand_conveyors}
\alias{expand_conveyors}
\title{Modify vars and consts list for conveyors}
\usage{
expand_conveyors(vars_xml, v_and_c, dt, dims_obj, inits_vector, vendor)
}
\arguments{
\item{vars_xml}{xml nodeset with all variables}

\item{v_and_c}{list to update if we find conveyors}

\item{dt}{simulation fractional time step}

\item{dims_obj}{dimensions object used to split arrayed stocks}

\item{inits_vector}{initialization overrides passed to compute_init_value}

\item{vendor}{isee or vensim (only supporting isee)}
}
\value{
modified list of vars and consts
}
\description{
Here we add a chain of "builtin stocks", along the lines of what is done
for DELAYN processing.  Each stock represents one slat in the conveyor, and
we connect them to each other and to the inflow(s), leak, and outflow(s).
For convenience, we add a variable with the conveyor name which is a vector
of all the slats.
}
