% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays.R
\name{expand_dimensions}
\alias{expand_dimensions}
\title{Expand a variable on it's dimensions.}
\usage{
expand_dimensions(var_name, dims_obj)
}
\arguments{
\item{var_name}{name of a variable (\emph{flow}, \emph{aux}, etc)}

\item{dims_obj}{object describing model dimensions}
}
\value{
list of expanded variable names, one for each combo of dimensions
}
\description{
If the variable has dimensions, return a list of new variables expanded on
the dimensions.
}
\details{
It is assumed that `dims_obj` contains
- `dictionary` which maps variables to lists of dimension names
- `global_dims` which maps dimension names to dimenion elements/categories
}
