% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveyors.R
\name{expand_one_conveyor}
\alias{expand_one_conveyor}
\title{Modify vars and consts list for a single conveyor stock}
\usage{
expand_one_conveyor(conveyor_name, cvy_parms)
}
\arguments{
\item{conveyor_name}{sanitized name of the conveyor stock}

\item{cvy_parms}{list of useful parameters}
}
\value{
list with builtin_stocks and variables to be added
}
\description{
This method is very specific to the Resilient model and not robust to all
forms of xmile!  For each conveyor, we generate a vector of "builtin stocks",
one for each slat.  These are lists with:
- name: built as "{conveyor_name}_slat_{slat_number}"
- equation: this is the equation that goes into differential calculation!
- initValue: initial value distributed over slats (see below)
}
\details{
Additionally, internal vector variables are used to keep track of per-slat
leakage, and those are summed into the scalar flow variables which may be
referenced downstream.  For the special case of downstream conveyors, the
vector version is used.


## Limitations
- we assume that all lengths are constant over the sim
  - handling variable transit times is much more complicated
- initial slat values hardcoded for the one conveyor with nonzero init
- right now, only partially vectorized (thus the mixture of tidyr and lists)
}
