% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveyors.R
\name{extract_conveyor_info}
\alias{extract_conveyor_info}
\title{Gather all the information we need on conveyors and connected flows.}
\usage{
extract_conveyor_info(vars_xml, vendor)
}
\arguments{
\item{vars_xml}{xml "variables" node to search under}

\item{vendor}{we only handle isee systems here}
}
\value{
list of dataframes describing conveyors and connected flows
}
\description{
Returns a list of dataframes
- conveyors with columns:
  - name: sanitized name (without any dimension info)
  - leak_exp: boolean indicating whether leaks are exponential or not
  - inflows: comma-separated list of inflow names (no dimensions)
  - outflows: comma-separated list of outflow names (no dimensions)
  - init_eqn: equation string for initialization
  - len_eqn: equation string for conveyor length
    - WARNING!  Assuming both eqn and len tags are unique within stock!
- flows (only inflows and outflows from above) with columns:
  - name: sanitized name (without any dimension info)
  - eqn: value of eqn tag directly under the flow
    - WARNING! we're not handling per-dimension eqn elements at this level
               we currently don't have any...
  - is_leak: boolean indicating whether the flow is a leak
  - spread_source: flow spreads from upstream to downstream conveyor slats
  - spread_even: flow spreads evenly into downstream conveyor slats
  - flows_from: sanitized name of conveyor this flows from or NA if none
  - flows_to: sanitized name of conveyor this flows to or NA if none
}
