% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_function_call}
\alias{extract_function_call}
\title{Extract one function call from a string.}
\usage{
extract_function_call(x, function_name)
}
\arguments{
\item{x}{A string of xmile.}

\item{function_name}{The name of the function to find.}
}
\value{
List with `match` being the full match and `args` just the part
  inside the parentheses.
}
\description{
Finds the given function name (not as part of another word), followed by
a matched set of parentheses.
}
