% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmile_helpers.R
\name{extract_parameters_one_constant}
\alias{extract_parameters_one_constant}
\title{Parse one constant (\code{<eqn>} node with numeric value)}
\usage{
extract_parameters_one_constant(node_eqn)
}
\arguments{
\item{node_eqn}{constant \code{eqn} node to extract info for}
}
\value{
A tibble
}
\description{
Pull out \emph{name}, \emph{dimensions}, \emph{element} for each dimensional
element, \emph{units}, and \emph{doc}.
}
\details{
\emph{WARNING} this does not handle multi dimensional arrays or integer
indexed arrays.  Only single dimensional arrays with named subscripts.
}
