% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCM_vizualization.R
\name{plot_all_conveyor_levels}
\alias{plot_all_conveyor_levels}
\title{Plot levels for all 8 of our conveyors}
\usage{
plot_all_conveyor_levels(results)
}
\arguments{
\item{results}{results returned from sd_simulate}
}
\description{
This just spits all plots out to file, no option to display them locally.
}
