% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCM_vizualization.R
\name{plot_conveyor_levels}
\alias{plot_conveyor_levels}
\title{Plot combined levels (sum of slat values) of one conveyor}
\usage{
plot_conveyor_levels(results, cname, fname = NA)
}
\arguments{
\item{results}{results returned from sd_simulate}

\item{cname}{name of conveyor to examine}

\item{fname}{if NA (default), plot is returned, else saved to this file}
}
\value{
plot unless fname specified
}
\description{
Plot combined levels (sum of slat values) of one conveyor
}
