% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveyors.R
\name{sd_conveyor_leak}
\alias{sd_conveyor_leak}
\title{Calculate conveyor leakage during simulation}
\usage{
sd_conveyor_leak(conveyor, fraction, is_exp)
}
\arguments{
\item{conveyor}{Conveyor stock value variable (array of slat values)}

\item{fraction}{Leak fraction (calculated from eqn element)}

\item{is_exp}{Exponential vs linear leakage}
}
\value{
The (positive) value contributing to conveyor incremental change.
}
\description{
For the exponential case, we leak a fraction of the slat value.
}
\details{
The linear case is more complicated.  You have to keep track of the original
value put into the conveyor (and where it was put in), then remove a fixed
fraction of that amount at every slat/timestep until that quantity exits the
conveyor.
}
