% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveyors.R
\name{sd_conveyor_outflow}
\alias{sd_conveyor_outflow}
\title{Calculate conveyor "normal" outflow during simulation}
\usage{
sd_conveyor_outflow(conveyor, leaks = 0)
}
\arguments{
\item{conveyor}{Conveyor stock value variable}

\item{leaks}{Leaks out of this conveyor (slat 1 only).}
}
\value{
The (positive) value contributing to conveyor incremental change.
}
\description{
Conveyors are broken into N slats numbered N down to 1.  Outflow comes from
slat 1 after subtracting all leakages.
}
\details{
Start with the full value of conveyor slat 1.  Divide by timestep so we take
the full value in one DT.  Then subtract out any leaks.  In all
cases, we are only concerned with slat 1.

NOTE: this assumes only one outflow per stock, which is true for RCM model!
      To handle other outflows, we need to collect and track flow priority.
}
