% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveyors.R
\name{sd_conveyor_slat_flow}
\alias{sd_conveyor_slat_flow}
\title{Calculate flow between slats during simulation}
\usage{
sd_conveyor_slat_flow(conveyor, ...)
}
\arguments{
\item{conveyor}{Conveyor stock value vector}

\item{...}{All other args assumed to be leak flow arrays from same conveyor}
}
\value{
The (positive) values contributing to slat incremental changes.
}
\description{
Start with the full value of conveyor slats.  Divide by timestep so we take
the full value in one DT.  Then subtract out any leak inputs.  Here the flow
with index i is the one going INTO slat i, so calculated from slats and leaks
with index i+1.
}
