% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveyors.R
\name{vec_array_equations}
\alias{vec_array_equations}
\title{Reformat equations with dimensional extensions}
\usage{
vec_array_equations(.data, eqn_col = eqn, cvy_parms)
}
\arguments{
\item{.data}{input data frame}

\item{eqn_col}{column to extract equations from (probably eqn or len)}

\item{cvy_parms}{conveyor parameters with vendor, dt, dmx}
}
\value{
modified data frame with
}
\description{
Sanitize variable names in equation then expand for all extensions.
}
